+++ crates/schematic/src/schema/renderers/template.rs
@@ -113,7 +113,7 @@ impl TemplateContext {
         }
     }
 
-    pub fn create_comment(&self, field: &Schema) -> String {
+    pub fn create_comment(&self, schema: &Schema) -> String {
         if !self.options.comments {
             return String::new();
         }
@@ -126,7 +126,44 @@ impl TemplateContext {
             lines.push(format!("{indent}{prefix}{line}"));
         };
 
-        if let Some(comment) = &field.description {
+        if let Some(comment) = &schema.description {
+            comment
+                .trim()
+                .split('\n')
+                .for_each(|c| push(c.trim().to_owned()));
+        }
+
+        if let Some(deprecated) = &schema.deprecated {
+            push(if deprecated.is_empty() {
+                "@deprecated".into()
+            } else {
+                format!("@deprecated {deprecated}")
+            });
+        }
+
+        if lines.is_empty() {
+            return String::new();
+        }
+
+        let mut out = lines.join("\n");
+        out.push('\n');
+        out
+    }
+
+    pub fn create_field_comment(&self, field: &SchemaField) -> String {
+        if !self.options.comments {
+            return String::new();
+        }
+
+        let mut lines = vec![];
+        let indent = self.indent();
+        let prefix = self.get_comment_prefix();
+
+        let mut push = |line: String| {
+            lines.push(format!("{indent}{prefix}{line}"));
+        };
+
+        if let Some(comment) = &field.comment {
             comment
                 .trim()
                 .split('\n')
@@ -154,12 +191,12 @@ impl TemplateContext {
         out
     }
 
-    pub fn create_field(&self, field: &Schema, property: String) -> String {
+    pub fn create_field(&self, field: &SchemaField, property: String) -> String {
         let key = self.get_stack_key();
 
         format!(
             "{}{}{}{property}",
-            self.create_comment(field),
+            self.create_field_comment(field),
             self.indent(),
             if self.options.comment_fields.contains(&key) {
                 self.get_comment_prefix()
@@ -196,7 +233,7 @@ impl TemplateContext {
         self.options.expand_fields.contains(key)
     }
 
-    pub fn is_hidden(&self, field: &Schema) -> bool {
+    pub fn is_hidden(&self, field: &SchemaField) -> bool {
         let key = self.get_stack_key();
 
         field.hidden || self.options.hide_fields.contains(&key)